Fixed synchronization bugs
